2/13 web / bbs mvc (redirect)

+++ src/net/bbs2/BbsDelete.java
@@ -12,19 +12,29 @@ public class BbsDelete implements CommandAction {
 
 		int num=Integer.parseInt(req.getParameter("num"));	// 글번호
 		String pageNum=req.getParameter("pageNum");	// 페이지번호
-
-		BoardDTO article=new BoardDTO();
-		article.setNum(Integer.parseInt(req.getParameter("num")));
-		article.setPasswd(req.getParameter("passwd").trim());
+		String ip=req.getRemoteAddr(); // 클라이언트 PC의 IP
+		boolean isAdmin=ip.equals("127.0.0.1") || ip.equals("0:0:0:0:0:0:0:1");
 		
-		BoardDAO dao=new BoardDAO();
-		int res=dao.deleteArticle(article);	// delete 수행
+		int res;
+		if(!isAdmin) {
+			res=-1;
+		}else {
+			// 관리자일 때만 삭제 기능
+			BoardDTO article=new BoardDTO();
+			article.setNum(Integer.parseInt(req.getParameter("num")));
+			article.setPasswd(req.getParameter("passwd").trim());
+			
+			BoardDAO dao=new BoardDAO();
+			res=dao.deleteArticle(article);	// delete 수행
+		}
 		
 		req.setAttribute("num", new Integer(num));
-		req.setAttribute("pageNum",new Integer(pageNum));
-		req.setAttribute("res",new Integer(res));	// query 성공여부
+		req.setAttribute("pageNum", new Integer(pageNum));
+		req.setAttribute("res", new Integer(res));	// query 성공여부
+		req.setAttribute("isAdmin", new Boolean(isAdmin));
+		req.setAttribute("action", "삭제");
 		
-		return "deleteProc.jsp";
+		return "bbsRedirect.jsp";	// deleteProc.jsp
 		
 	} // requestPro() end
